requirements.txt
matplotlib
pandas
xhtml2pdf
celery
django-celery-beat
openpyxl
plotly
